+++ modules/hanna-react/CHANGELOG.md
@@ -6,6 +6,23 @@
 - `BasicTable`:
   - feat: Add props `wrapperProps`
   - feat: Export `Table*` types (`TableCell`, `TableRow`, etc.)
+- feat: Add component `AutosuggestSearch`
+- `SearchInput`:
+  - feat: Add prop `onSubmit` firing on `Enter` keypress
+  - feat: `onButtonClick` receives `value` as an argument
+  - feat: Use `onSubmit` as fallback for `onButtonClick`
+  - feat: Add prop `button` to toggle the button, ueses`onSubmit` as fallback
+  - fix: Minor prop and attribute assignment cleanup
+- `SiteSearchInput`:
+  - feat: `onButtonClick` and `onSubmit` receive `value` as their argument
+  - feat: Add prop `button` to toggle the button, ueses`onSubmit` as fallback
+  - fix: Minor prop and attribute assignment cleanup
+- `SiteSearchAutocomplete`:
+  - fix: Add translation prop `buttonText` — (optional, but logs warning in
+    dev if missing.)
+  - fix: Spread `wrapperProps` on the wrapping `<div/>` element not the
+    contained `FormField`
+  - feat: Deprecate this component, in favor of `AutosuggestSearch`
 - feat: Add `useDomid` to `utils`
 
 ## 0.10.103